Stabilize markdown block command ranges
