+++ app/components/editor/MarkdownEditor.tsx
@@ -23,7 +23,6 @@ import { TaskItem } from '@tiptap/extension-task-item';
 import { TableKit } from '@tiptap/extension-table';
 import { CodeBlock } from '@tiptap/extension-code-block';
 import { Suggestion, type SuggestionProps } from '@tiptap/suggestion';
-import { Node as ProseMirrorNode } from '@tiptap/pm/model';
 import { Plugin, PluginKey } from '@tiptap/pm/state';
 import { Decoration, DecorationSet } from '@tiptap/pm/view';
 import {
@@ -81,8 +80,20 @@ import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '@/comp
 import { SafeMarkdownImage } from '@/app/components/shared/SafeMarkdownImage';
 import {
   clampEditorRangeToDoc,
+  getSlashCommandDeletionRange,
+  isEditorRangeInsideDoc,
   isEditorPositionInsideDoc,
 } from '@/app/lib/editor/prosemirror-ranges';
+import {
+  createCurrentBlockCommandTarget,
+  createInsertedBlockCommandTarget,
+  getBlockDropInsertPosition,
+  getBlockInsertButtonPosition,
+  moveReorderableBlock,
+  type BlockControlPosition,
+  type BlockInsertPlacement,
+  type ReorderableBlockRange,
+} from '@/app/lib/editor/reorderable-blocks';
 import { createInlineColorRegex, isColorCode } from '@/app/lib/markdown/color-code';
 import { makeLinkPreviewImageAlt, parseLinkPreviewImageAlt } from '@/app/lib/markdown/link-preview-markdown';
 import {
@@ -204,26 +215,6 @@ type BlockCommandMenuState = {
     top: number;
     width: number;
   };
-};
-
-type BlockInsertPlacement = 'above' | 'below';
-
-type ReorderableBlockKind = 'topLevel' | 'listItem';
-
-type ReorderableBlockRange = {
-  depth: number;
-  from: number;
-  kind: ReorderableBlockKind;
-  node: ProseMirrorNode;
-  parentFrom: number;
-  parentTo: number;
-  to: number;
-};
-
-type BlockControlPosition = {
-  blockRange: ReorderableBlockRange;
-  menuRange: Range;
-  top: number;
 };
 
 type ImageDialogSeed = {
@@ -435,13 +426,33 @@ const ColorSwatchDecorations = Extension.create({
 
 function runAfterSlashDelete({ editor, range }: SlashCommandContext) {
   const chain = editor.chain().focus();
-  const safeRange = clampEditorRangeToDoc(editor, range);
-
-  if (!safeRange || safeRange.from === safeRange.to) {
+  if (!isEditorRangeInsideDoc(editor, range)) {
     return chain;
   }
 
-  return chain.deleteRange(safeRange);
+  if (range.from === range.to) {
+    return chain.setTextSelection(range.from);
+  }
+
+  const slashRange = getSlashCommandDeletionRange(editor, range);
+  return slashRange ? chain.deleteRange(slashRange) : chain.setTextSelection(range.from);
+}
+
+function prepareCommandDialogInsertionRange(editor: Editor, range: Range): Range | null {
+  if (!isEditorRangeInsideDoc(editor, range)) {
+    return null;
+  }
+
+  const slashRange = getSlashCommandDeletionRange(editor, range);
+  const chain = editor.chain().focus();
+
+  if (slashRange) {
+    chain.deleteRange(slashRange).run();
+  } else {
+    chain.setTextSelection(range.from).run();
+  }
+
+  return { from: range.from, to: range.from };
 }
 
 const SLASH_COMMAND_DEFINITIONS: SlashCommandDefinition[] = [
@@ -524,12 +535,12 @@ const SLASH_COMMAND_DEFINITIONS: SlashCommandDefinition[] = [
 
       const src = window.prompt(labels.imageSrcPrompt);
       if (!src?.trim()) {
-        editor.chain().focus().deleteRange(range).run();
+        runAfterSlashDelete({ actions, editor, labels, range }).run();
         return;
       }
 
       const alt = window.prompt(labels.imageAltPrompt) || '';
-      editor.chain().focus().deleteRange(range).setImage({ src: src.trim(), alt: alt.trim() }).run();
+      runAfterSlashDelete({ actions, editor, labels, range }).setImage({ src: src.trim(), alt: alt.trim() }).run();
     },
   },
   {
@@ -922,162 +933,9 @@ function createSlashCommandLabels(t: (key: string) => string): SlashCommandLabel
   };
 }
 
-function findActiveTextblockDepth(editor: Editor): number | null {
-  const { $from } = editor.state.selection;
-
-  for (let depth = $from.depth; depth > 0; depth -= 1) {
-    if ($from.node(depth).isTextblock) return depth;
-  }
-
-  return null;
-}
-
-function getTopLevelBlockRangeAt(editor: Editor, position: number): ReorderableBlockRange | null {
-  const docEnd = editor.state.doc.content.size;
-  const safePosition = Math.max(0, Math.min(position, docEnd));
-  let range: ReorderableBlockRange | null = null;
-
-  editor.state.doc.forEach((node, offset) => {
-    if (range) return;
-
-    const from = offset;
-    const to = offset + node.nodeSize;
-    const isInsideNode = safePosition >= from && safePosition < to;
-    const isAtDocumentEnd = safePosition === docEnd && safePosition === to;
-
-    if (isInsideNode || isAtDocumentEnd) {
-      range = {
-        depth: 1,
-        from,
-        kind: 'topLevel',
-        node,
-        parentFrom: 0,
-        parentTo: docEnd,
-        to,
-      };
-    }
-  });
-
-  return range;
-}
-
-function getListItemBlockRangeAt(
-  editor: Editor,
-  position: number,
-  requiredParent?: Pick<ReorderableBlockRange, 'parentFrom' | 'parentTo'>,
-): ReorderableBlockRange | null {
-  const doc = editor.state.doc;
-  const docEnd = doc.content.size;
-  if (docEnd <= 0) return null;
-
-  const safePosition = Math.max(0, Math.min(position, docEnd));
-  const resolvePosition = Math.max(0, Math.min(safePosition, docEnd - 1));
-  const $position = doc.resolve(resolvePosition);
-
-  for (let depth = $position.depth; depth > 0; depth -= 1) {
-    const node = $position.node(depth);
-    if (node.type.name !== 'listItem') continue;
-
-    const parentDepth = depth - 1;
-    const parentNode = $position.node(parentDepth);
-    if (parentNode.type.name !== 'bulletList' && parentNode.type.name !== 'orderedList' && parentNode.type.name !== 'taskList') {
-      continue;
-    }
-
-    const parentFrom = $position.start(parentDepth);
-    const parentTo = $position.end(parentDepth);
-    if (requiredParent && (parentFrom !== requiredParent.parentFrom || parentTo !== requiredParent.parentTo)) {
-      continue;
-    }
-
-    return {
-      depth,
-      from: $position.before(depth),
-      kind: 'listItem',
-      node,
-      parentFrom,
-      parentTo,
-      to: $position.after(depth),
-    };
-  }
-
-  return null;
-}
-
-function getReorderableBlockRangeAt(
-  editor: Editor,
-  position: number,
-  source?: ReorderableBlockRange,
-): ReorderableBlockRange | null {
-  if (source?.kind === 'listItem') {
-    return getListItemBlockRangeAt(editor, position, source);
-  }
-
-  if (source?.kind === 'topLevel') {
-    return getTopLevelBlockRangeAt(editor, position);
-  }
-
-  return getListItemBlockRangeAt(editor, position) ?? getTopLevelBlockRangeAt(editor, position);
-}
-
-function createEmptyListItemNode(editor: Editor, source: ReorderableBlockRange) {
-  const paragraph = editor.schema.nodes.paragraph.create();
-  return source.node.type.createAndFill(null, paragraph) ?? source.node.type.create(null, paragraph);
-}
-
-function createInsertedBlockCommandTarget(
-  editor: Editor,
-  placement: BlockInsertPlacement,
-  blockRange?: ReorderableBlockRange,
-): Range | null {
-  if (!editor.isEditable || editor.isActive('codeBlock')) return null;
-
-  if (blockRange) {
-    const insertPosition = placement === 'above' ? blockRange.from : blockRange.to;
-    const isListItem = blockRange.kind === 'listItem';
-    const cursorPosition = insertPosition + (isListItem ? 2 : 1);
-    const content = isListItem ? createEmptyListItemNode(editor, blockRange) : { type: 'paragraph' };
-
-    editor.chain().focus().insertContentAt(insertPosition, content).setTextSelection(cursorPosition).run();
-
-    return { from: cursorPosition, to: cursorPosition };
-  }
-
-  const { $from } = editor.state.selection;
-  const textblockDepth = findActiveTextblockDepth(editor);
-  if (!textblockDepth) return null;
-
-  const topLevelDepth = $from.depth >= 1 ? 1 : textblockDepth;
-  const insertPosition = placement === 'above' ? $from.before(topLevelDepth) : $from.after(topLevelDepth);
-  const cursorPosition = insertPosition + 1;
-  editor
-    .chain()
-    .focus()
-    .insertContentAt(insertPosition, { type: 'paragraph' })
-    .setTextSelection(cursorPosition)
-    .run();
-
-  return { from: cursorPosition, to: cursorPosition };
-}
-
-function createCurrentBlockCommandTarget(editor: Editor, menuRange?: Range): Range | null {
-  if (!editor.isEditable || editor.isActive('codeBlock')) return null;
-
-  if (menuRange) {
-    editor.chain().focus().setTextSelection(menuRange.from).run();
-    return menuRange;
-  }
-
-  const { $from } = editor.state.selection;
-  const textblockDepth = findActiveTextblockDepth(editor);
-  if (!textblockDepth) return null;
-
-  const position = $from.start(textblockDepth);
-  editor.chain().focus().setTextSelection(position).run();
-  return { from: position, to: position };
-}
-
 function createBlockCommandMenuState(editor: Editor, range: Range): BlockCommandMenuState | null {
+  if (!isEditorRangeInsideDoc(editor, range)) return null;
+
   try {
     const coords = editor.view.coordsAtPos(range.from);
     return {
@@ -1087,96 +945,6 @@ function createBlockCommandMenuState(editor: Editor, range: Range): BlockCommand
     };
   } catch {
     return null;
-  }
-}
-
-function getBlockInsertButtonPosition(editor: Editor, container: HTMLDivElement): BlockControlPosition | null {
-  if (!editor.isEditable || editor.isActive('codeBlock')) return null;
-
-  const { $from } = editor.state.selection;
-  const textblockDepth = findActiveTextblockDepth(editor);
-  if (!textblockDepth) return null;
-
-  const blockRange = getReorderableBlockRangeAt(editor, editor.state.selection.from);
-  if (!blockRange) return null;
-
-  const blockDom = editor.view.nodeDOM(blockRange.from);
-  const containerRect = container.getBoundingClientRect();
-  const menuPosition = $from.start(textblockDepth);
-  const menuRange = { from: menuPosition, to: menuPosition };
-
-  if (blockDom instanceof HTMLElement) {
-    const blockRect = blockDom.getBoundingClientRect();
-    return {
-      blockRange,
-      menuRange,
-      top: Math.max(6, blockRect.top - containerRect.top + container.scrollTop + (blockRect.height / 2) - 12),
-    };
-  }
-
-  const positionForCoords = Math.min(blockRange.from + 1, editor.state.doc.content.size);
-  const coords = editor.view.coordsAtPos(positionForCoords);
-
-  return {
-    blockRange,
-    menuRange,
-    top: Math.max(6, coords.top - containerRect.top + container.scrollTop),
-  };
-}
-
-function getBlockDropInsertPosition(
-  editor: Editor,
-  event: DragEvent,
-  source: ReorderableBlockRange,
-): number | null {
-  const positionAtCoords = editor.view.posAtCoords({
-    left: event.clientX,
-    top: event.clientY,
-  });
-
-  if (!positionAtCoords) {
-    return source.kind === 'topLevel' ? editor.state.doc.content.size : null;
-  }
-
-  const target = getReorderableBlockRangeAt(editor, positionAtCoords.pos, source);
-  if (!target) {
-    return null;
-  }
-
-  const targetDom = editor.view.nodeDOM(target.from);
-  const targetRect = targetDom instanceof HTMLElement ? targetDom.getBoundingClientRect() : null;
-  const insertPosition = targetRect
-    ? event.clientY < targetRect.top + targetRect.height / 2
-      ? target.from
-      : target.to
-    : positionAtCoords.pos <= target.from
-      ? target.from
-      : target.to;
-
-  if (insertPosition >= source.from && insertPosition <= source.to) {
-    return null;
-  }
-
-  return insertPosition;
-}
-
-function moveReorderableBlock(editor: Editor, source: ReorderableBlockRange, insertPosition: number) {
-  const sourceSize = source.to - source.from;
-  const adjustedInsertPosition = insertPosition > source.from ? insertPosition - sourceSize : insertPosition;
-
-  if (adjustedInsertPosition === source.from) return;
-
-  try {
-    const transaction = editor.state.tr
-      .delete(source.from, source.to)
-      .insert(adjustedInsertPosition, source.node)
-      .scrollIntoView();
-
-    editor.view.dispatch(transaction);
-    editor.commands.focus();
-  } catch {
-    // Ignore invalid drops, for example when the browser reports coordinates
-    // outside the reorderable parent list.
   }
 }
 
@@ -2730,6 +2498,7 @@ function RichMarkdownEditor({
   const latestValueRef = useRef(value);
   const acceptedExternalValueRef = useRef(value);
   const applyingExternalValueRef = useRef(false);
+  const pendingBlockCommandMenuFrameRef = useRef<number | null>(null);
   const scrollContainerRef = useRef<HTMLDivElement>(null);
   const [tableDialogOpen, setTableDialogOpen] = useState(false);
   const [imageDialogOpen, setImageDialogOpen] = useState(false);
@@ -2743,14 +2512,8 @@ function RichMarkdownEditor({
     setImageDialogOpen(open);
   }, []);
   const openImageDialogFromSlash = useCallback((slashEditor: Editor, range: Range) => {
-    const safeRange = clampEditorRangeToDoc(slashEditor, range);
-    const insertPosition = safeRange?.from ?? slashEditor.state.selection.from;
-
-    if (safeRange && safeRange.from !== safeRange.to) {
-      slashEditor.chain().focus().deleteRange(safeRange).run();
-    } else {
-      slashEditor.chain().focus().setTextSelection(insertPosition).run();
-    }
+    const insertionRange = prepareCommandDialogInsertionRange(slashEditor, range);
+    const insertPosition = insertionRange?.from ?? slashEditor.state.selection.from;
 
     setImageDialogSeed((current) => ({
       id: current.id + 1,
@@ -2759,13 +2522,7 @@ function RichMarkdownEditor({
     setImageDialogOpen(true);
   }, []);
   const openTableDialogFromSlash = useCallback((slashEditor: Editor, range: Range) => {
-    const safeRange = clampEditorRangeToDoc(slashEditor, range);
-
-    if (safeRange && safeRange.from !== safeRange.to) {
-      slashEditor.chain().focus().deleteRange(safeRange).run();
-    } else if (safeRange) {
-      slashEditor.chain().focus().setTextSelection(safeRange.from).run();
-    }
+    prepareCommandDialogInsertionRange(slashEditor, range);
 
     setTableDialogOpen(true);
   }, []);
@@ -2807,18 +2564,33 @@ function RichMarkdownEditor({
     setTableDialogOpen(false);
   }, [editor]);
 
-  const closeBlockCommandMenu = useCallback(() => {
-    setBlockCommandMenu(null);
+  const cancelPendingBlockCommandMenu = useCallback(() => {
+    if (pendingBlockCommandMenuFrameRef.current === null) return;
+    window.cancelAnimationFrame(pendingBlockCommandMenuFrameRef.current);
+    pendingBlockCommandMenuFrameRef.current = null;
   }, []);
 
+  const closeBlockCommandMenu = useCallback(() => {
+    cancelPendingBlockCommandMenu();
+    setBlockCommandMenu(null);
+  }, [cancelPendingBlockCommandMenu]);
+
   const openBlockCommandMenuAtRange = useCallback((blockEditor: Editor, range: Range) => {
-    window.requestAnimationFrame(() => {
+    cancelPendingBlockCommandMenu();
+    setBlockCommandMenu(null);
+
+    if (!isEditorRangeInsideDoc(blockEditor, range)) return;
+
+    pendingBlockCommandMenuFrameRef.current = window.requestAnimationFrame(() => {
+      pendingBlockCommandMenuFrameRef.current = null;
+      if (!blockEditor.isEditable || !isEditorRangeInsideDoc(blockEditor, range)) return;
+
       const menuState = createBlockCommandMenuState(blockEditor, range);
       if (menuState) {
         setBlockCommandMenu(menuState);
       }
     });
-  }, []);
+  }, [cancelPendingBlockCommandMenu]);
 
   const openInsertedBlockCommandMenu = useCallback((
     blockEditor: Editor,
@@ -2837,6 +2609,10 @@ function RichMarkdownEditor({
 
     openBlockCommandMenuAtRange(blockEditor, range);
   }, [openBlockCommandMenuAtRange]);
+
+  useEffect(() => () => {
+    cancelPendingBlockCommandMenu();
+  }, [cancelPendingBlockCommandMenu]);
 
   useEffect(() => {
     if (!markdownEditor) return;
@@ -2891,11 +2667,11 @@ function RichMarkdownEditor({
     if (!readOnly) return undefined;
 
     const frame = window.requestAnimationFrame(() => {
-      setBlockCommandMenu(null);
+      closeBlockCommandMenu();
     });
 
     return () => window.cancelAnimationFrame(frame);
-  }, [readOnly]);
+  }, [closeBlockCommandMenu, readOnly]);
 
   return (
     <div className="flex h-full min-h-0 flex-col overflow-hidden bg-background">
